Refactor: move HttpStack-related classes to separate namespace
The moved classes have a high cohesion and are thus moved to their own package. Also, the HttpResponseStackBuilder was renamed to HttpStackBuilder as it builds both, HttpErrorStack and HttpResponseStack.
