+++ src/regression/assertion/assertion.ts
@@ -1,4 +1,4 @@
-import {HttpResponseStack} from '../http-response-stack';
+import {HttpResponseStack} from '../stack/http-response-stack';
 import {AssertionResult} from './assertion-result';
 
 export interface Assertion {
