+++ src/regression/assertion/assertions.ts
@@ -1,5 +1,5 @@
 import {Assertion} from './assertion';
-import {HttpResponseStack} from '../http-response-stack';
+import {HttpResponseStack} from '../stack/http-response-stack';
 import {AssertionResult} from './assertion-result';
 
 export class Assertions {
